# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProxyAdminPanel",
    "description": "X-UI style proxy management system with BFF architecture",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/outbounds": {
      "get": {
        "tags": [
          "Outbounds"
        ],
        "summary": "Get All Outbounds",
        "description": "Get all outbound configurations.",
        "operationId": "get_all_outbounds_api_outbounds_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OutboundResponse"
                  },
                  "type": "array",
                  "title": "Response Get All Outbounds Api Outbounds Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Outbounds"
        ],
        "summary": "Create Outbound",
        "description": "Create new outbound configuration.",
        "operationId": "create_outbound_api_outbounds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutboundCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutboundResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/outbounds/{outbound_id}": {
      "get": {
        "tags": [
          "Outbounds"
        ],
        "summary": "Get Outbound",
        "description": "Get outbound by ID.",
        "operationId": "get_outbound_api_outbounds__outbound_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outbound_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Outbound Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutboundResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Outbounds"
        ],
        "summary": "Update Outbound",
        "description": "Update existing outbound configuration.",
        "operationId": "update_outbound_api_outbounds__outbound_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outbound_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Outbound Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutboundUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutboundResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Outbounds"
        ],
        "summary": "Delete Outbound",
        "description": "Delete outbound configuration.",
        "operationId": "delete_outbound_api_outbounds__outbound_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outbound_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Outbound Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/outbounds/scan": {
      "post": {
        "tags": [
          "Outbounds"
        ],
        "summary": "Scan Local Interfaces",
        "description": "One-click scan: Automatically create outbounds for all local network interfaces.\nThis is a key feature for quickly setting up direct outbounds.",
        "operationId": "scan_local_interfaces_api_outbounds_scan_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/game-inventory/": {
      "get": {
        "tags": [
          "game-inventory"
        ],
        "summary": "Get All Inventories",
        "description": "Get inventory for all games.\nShows available IPs for each game/rule.",
        "operationId": "get_all_inventories_api_game_inventory__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameInventoryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/game-inventory/{rule_id}": {
      "get": {
        "tags": [
          "game-inventory"
        ],
        "summary": "Get Game Inventory",
        "description": "Get inventory for a specific game/rule.",
        "operationId": "get_game_inventory_api_game_inventory__rule_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameInventory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game-inventory/outbound/{outbound_id}/usage": {
      "get": {
        "tags": [
          "game-inventory"
        ],
        "summary": "Get Outbound Usage",
        "description": "Get detailed usage information for a specific outbound/IP.\nShows which games are using this IP and how many users.",
        "operationId": "get_outbound_usage_api_game_inventory_outbound__outbound_id__usage_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "outbound_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Outbound Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GameInventory": {
        "properties": {
          "rule_id": {
            "type": "integer",
            "title": "Rule Id"
          },
          "rule_name": {
            "type": "string",
            "title": "Rule Name"
          },
          "total_ips": {
            "type": "integer",
            "title": "Total Ips"
          },
          "available_ips": {
            "type": "integer",
            "title": "Available Ips"
          },
          "used_ips": {
            "type": "integer",
            "title": "Used Ips"
          }
        },
        "type": "object",
        "required": [
          "rule_id",
          "rule_name",
          "total_ips",
          "available_ips",
          "used_ips"
        ],
        "title": "GameInventory",
        "description": "Inventory information for a game (rule)"
      },
      "GameInventoryResponse": {
        "properties": {
          "games": {
            "items": {
              "$ref": "#/components/schemas/GameInventory"
            },
            "type": "array",
            "title": "Games"
          },
          "total_outbounds": {
            "type": "integer",
            "title": "Total Outbounds"
          }
        },
        "type": "object",
        "required": [
          "games",
          "total_outbounds"
        ],
        "title": "GameInventoryResponse",
        "description": "Response with all game inventories"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OutboundCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "default": "direct"
          },
          "local_interface_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Interface Ip"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          },
          "max_users": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Users",
            "default": 10
          },
          "config": {
            "type": "object",
            "title": "Config"
          },
          "is_auto_generated": {
            "type": "boolean",
            "title": "Is Auto Generated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "OutboundCreate"
      },
      "OutboundResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "default": "direct"
          },
          "local_interface_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Interface Ip"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          },
          "max_users": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Users",
            "default": 10
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "config": {
            "type": "object",
            "title": "Config"
          },
          "is_auto_generated": {
            "type": "boolean",
            "title": "Is Auto Generated"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "config",
          "is_auto_generated",
          "created_at",
          "updated_at"
        ],
        "title": "OutboundResponse"
      },
      "OutboundUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "protocol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol"
          },
          "config": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "local_interface_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Interface Ip"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          },
          "max_users": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Users"
          }
        },
        "type": "object",
        "title": "OutboundUpdate"
      },
...
      "SuccessResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "SuccessResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}